+++ SdkGenerator/SdkGenerator/Program.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using CommandLine;
 using Newtonsoft.Json;
-using SdkGenerator.Project;
+using SdkGenerator.Sdks;
 
 namespace SdkGenerator;
