+++ SdkGenerator/SdkGenerator/Sdks/JavaSdk.cs
@@ -4,10 +4,9 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using SdkGenerator.Project;
 using SdkGenerator.Schema;
 
-namespace SdkGenerator;
+namespace SdkGenerator.Sdks;
 
 public static class JavaSdk
 {
